feat : Memory Nav style코드 분리
