+++ src/pages/MemoryPage/components/MemoryNav.tsx
@@ -1,77 +1,8 @@
-import styled, { css } from 'styled-components';
 import { useEffect, useState } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { FormControl, MenuItem, Select, SelectChangeEvent } from '@mui/material';
-import { BlackLink, IconButton } from '../../../styles/CommonStyle';
-
-const hoverCss = css`
-  .nav_pageTittle {
-    color: ${(props) => props.theme.color.main.orange};
-  }
-  .navIcon {
-    filter: invert(67%) sepia(15%) saturate(5627%) hue-rotate(336deg) brightness(103%)
-      contrast(101%);
-  }
-`;
-
-const NavBarContainer = styled.nav`
-  display: flex;
-  width: 80vw;
-  height: 132px;
-  margin: 0 auto;
-  background-color: ${(props) => props.theme.color.grayScale.white};
-  align-items: center;
-  color: black;
-  border-bottom: 2px solid #ebebeb;
-  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-  position: absolute;
-  top: 560px;
-  left: 10vw;
-  border-radius: 16px;
-  img {
-    cursor: pointer;
-  }
-
-  .navbox {
-    display: flex;
-    align-items: center;
-    justify-content: space-around;
-    border-right: 2px solid #dddddd;
-    height: 24px;
-    flex: 1;
-
-    &:hover {
-      ${hoverCss}
-    }
-
-    &.active {
-      ${hoverCss}
-    }
-  }
-
-  .nav_title {
-    font-family: ${(props) => props.theme.font.family.pretendard_bold};
-    font-size: 24px;
-    flex: 1;
-    text-align: center;
-    margin-left: 3.75vw;
-  }
-
-  .nav_pageTittle {
-    color: ${(props) => props.theme.color.grayScale.gray};
-    font-size: 16px;
-    font-family: ${(props) => props.theme.font.family.pretendard_medium};
-    margin-left: 1vw;
-  }
-
-  .nav_form {
-    flex: 2;
-    display: flex;
-    align-items: center;
-    justify-content: flex-end;
-    margin-right: 3.75vw;
-  }
-`;
+import { BlackLink, IconButton } from '../../../components/CommonStyle';
+import * as S from './style/MemoryNavStyle';
 
 function MemoryNav() {
   const [activeNum, setActiveNum] = useState(1);
@@ -104,7 +35,7 @@ function MemoryNav() {
   }, [location]);
 
   return (
-    <NavBarContainer>
+    <S.NavBarContainer>
       <div className="nav_title">추억앨범 서비스</div>
       <nav className={activeNum === 1 ? 'navbox active' : 'navbox'} onClick={handleNavToOne}>
         <BlackLink to="sharedAlbum">
@@ -170,7 +101,7 @@ function MemoryNav() {
           <img src="/img/앨범쓰기.svg" alt="앨범쓰기" />
         </IconButton>
       </div>
-    </NavBarContainer>
+    </S.NavBarContainer>
   );
 }
